pyproject.toml
[tool.poetry]
name = "planches-telegram-bot"
version = "0.1.0"
description = "Telegram bot for https://planches.arnalo.ch/"
authors = ["Stache <stache@stache.cat>"]
license = "GPL-3.0 license"

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.12.0"
isort = "^5.12.0"
pre-commit = "^3.0.1"
pylint = "^2.15.10"
bump2version = "^1.0.1"
python-telegram-bot = "^20.0"
peewee = "^3.15.4"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
# When imports are broken into multi-line, use the "Vertical Hanging Indent" layout.
multi_line_output = 3

# Always add a trailing comma to import lists (default: False).
include_trailing_comma = true

# Always put imports lists into vertical mode (0 = none allowed on first line)
force_grid_wrap = 0

# When multi-lining imports, use parentheses for line-continuation instead of default \.
use_parentheses = true

# Max import line length.
line_length = 120

# Regardless of what follows the imports, force 2 blank lines after the import list
lines_after_imports = 2

# Insert 2 blank lines between each section
lines_between_sections = 2

# Alphabetical sort in sections (inside a line or in ())
force_alphabetical_sort_within_sections = true

# Sort by lexicographical
lexicographical = true

# Put all from before import
from_first = true

ensure_newline_before_comments = true
